+++ R/risks.R
@@ -16,7 +16,8 @@
 #'
 #' \code{\link[risks]{confint.robpoisson}},
 #' \code{\link[risks]{confint.duplicate}},
-#' \code{\link[risks]{confint.margstd}}: Confidence intervals. (Standard
+#' \code{\link[risks]{confint.margstd_boot}},
+#' \code{\link[risks]{confint.margstd_delta}}: Confidence intervals. (Standard
 #'   confidence intervals for generalized linear models are used for other
 #'   models.)
 #'