+++ tests/testthat/test-coefficients.R
@@ -19,8 +19,10 @@ test_that("RD coefficients for receptorLow are the same", {
                                  approach = "glm_cem"))
   rd_cem_startp  <- coef(riskdiff(formula = death ~ receptor, data = dat,
                                  approach = "glm_cem_startp"))
-  rd_margstd    <- coef(riskdiff(formula = death ~ receptor, data = dat,
-                                 approach = "margstd"))
+  rd_margstd_boot <- coef(riskdiff(formula = death ~ receptor, data = dat,
+                                   approach = "margstd_boot"))
+  rd_margstd_delta <- coef(riskdiff(formula = death ~ receptor, data = dat,
+                                    approach = "margstd_delta"))
   rd_mh <- rr_rd_mantel_haenszel(data = dat, exposure = receptor,
                                  outcome = death,
                                  estimand = "rd")$estimate[1]
@@ -37,15 +39,19 @@ test_that("RD coefficients for receptorLow are the same", {
                tolerance = tol)
   expect_equal((23/48) - (31/144), as.numeric(rd_cem_startp["receptorLow"]),
                tolerance = tol)
-  expect_equal((23/48) - (31/144), as.numeric(rd_margstd["receptorLow"]),
+  expect_equal((23/48) - (31/144), as.numeric(rd_margstd_boot["receptorLow"]),
+               tolerance = tol)
+  expect_equal((23/48) - (31/144), as.numeric(rd_margstd_delta["receptorLow"]),
                tolerance = tol)
   expect_equal((23/48) - (31/144), rd_mh,    tolerance = tol)
   expect_equal(rd_glm, rd_glm_startp, tolerance = 0.03)
   expect_equal(rd_glm, rd_robpoisson, tolerance = 0.03)
   expect_equal(rd_glm, rd_cem, tolerance = 0.01)
   expect_equal(rd_glm, rd_cem_startp, tolerance = 0.01)
-  expect_equal(rd_glm["receptorLow"], rd_margstd["receptorLow"],
+  expect_equal(rd_glm["receptorLow"], rd_margstd_boot["receptorLow"],
                tolerance = 0.03)
+  expect_equal(rd_glm["receptorLow"], rd_margstd_delta["receptorLow"],
+               tolerance = 0.01)
 })
 
 test_that("RR coefficients for receptorLow are the same", {
@@ -71,8 +77,10 @@ test_that("RR coefficients for receptorLow are the same", {
                                   approach = "glm_cem"))
   rr_cem_startp  <- coef(riskratio(formula = death ~ receptor, data = dat,
                                   approach = "glm_cem_startp"))
-  rr_margstd    <- coef(riskratio(formula = death ~ receptor, data = dat,
-                                  approach = "margstd"))
+  rr_margstd_boot <- coef(riskratio(formula = death ~ receptor, data = dat,
+                                    approach = "margstd_boot"))
+  rr_margstd_delta <- coef(riskratio(formula = death ~ receptor, data = dat,
+                                     approach = "margstd_delta"))
   rr_mh <- rr_rd_mantel_haenszel(data = dat, exposure = receptor,
                                  outcome = death,
                                  estimand = "rr")$estimate[1]
@@ -91,7 +99,9 @@ test_that("RR coefficients for receptorLow are the same", {
                tolerance = tol)
   expect_equal(log((23/48) / (31/144)), as.numeric(rr_cem_startp["receptorLow"]),
                tolerance = tol)
-  expect_equal(log((23/48) / (31/144)), as.numeric(rr_margstd["receptorLow"]),
+  expect_equal(log((23/48) / (31/144)), as.numeric(rr_margstd_boot["receptorLow"]),
+               tolerance = tol)
+  expect_equal(log((23/48) / (31/144)), as.numeric(rr_margstd_delta["receptorLow"]),
                tolerance = tol)
   expect_equal(log((23/48) / (31/144)), rr_mh,
                tolerance = tol)
@@ -101,8 +111,12 @@ test_that("RR coefficients for receptorLow are the same", {
   expect_equal(rr_glm, rr_robpoisson, tolerance = 0.03)
   expect_equal(rr_glm, rr_cem, tolerance = 0.01)
   expect_equal(rr_glm, rr_cem_startp, tolerance = 0.01)
-  expect_equal(rr_glm["receptorLow"], rr_margstd["receptorLow"],
+  expect_equal(rr_glm["receptorLow"], rr_margstd_boot["receptorLow"],
                tolerance = 0.03)
+  expect_equal(rr_glm["receptorLow"], rr_margstd_delta["receptorLow"],
+               tolerance = 0.01)
+})
+
 test_that("Continuous and implicit binary variables pass in marg std", {
   dat <- data.frame(
     death    = c(rep(1, 54), rep(0, 138)),